+++ src/InputOutputManager.hpp
@@ -77,7 +77,7 @@ class TemporalPointProbe {
 
         ////////// CONSTRUCTORS & DESTRUCTOR //////////
         TemporalPointProbe();					/// Default constructor
-        TemporalPointProbe(const int &x_position_, const int &y_position_, const int &z_position_, const std::string output_file_name_, ComputationalDomain* mesh_, ParallelTopology* topo_);			     /// Parametrized constructor
+        TemporalPointProbe(const double &x_position_, const double &y_position_, const double &z_position_, const std::string output_file_name_, ComputationalDomain* mesh_, ParallelTopology* topo_);			     /// Parametrized constructor
         virtual ~TemporalPointProbe();				/// Destructor
 
 	////////// GET FUNCTIONS //////////
